requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
pypdf
transformers>=4.46
accelerate
torch
einops
